+++ src/main/java/com/nnk/springboot/services/RuleNameImplService.java
@@ -1,12 +1,15 @@
 package com.nnk.springboot.services;
 
+import com.nnk.springboot.domain.Rating;
 import com.nnk.springboot.domain.RuleName;
+import com.nnk.springboot.exceptions.RecordNotFoundException;
 import com.nnk.springboot.repositories.RuleNameRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -15,17 +18,25 @@ public class RuleNameImplService implements IRuleNameService {
     @Autowired
     RuleNameRepository ruleNameRepository;
 
-    public List<RuleName> getAllRuleNames() {
+    public List<RuleName> findAllRuleNames() {
         return ruleNameRepository.findAll();
     }
 
     @Override
-    public RuleName getRuleNameById(Integer id) {
-        return ruleNameRepository.getOne(id);
+    public RuleName findRuleNameById(Integer id) throws RecordNotFoundException {
+        Optional<RuleName> ruleName = ruleNameRepository.findById(id);
+
+        if(ruleName.isPresent()) {
+            return ruleName.get();
+        } else {
+            throw new RecordNotFoundException("No record exist for given id");
+        }
+
+        //return ruleNameRepository.getOne(id);
     }
 
     @Override
-    public RuleName saveRuleName(RuleName RuleName) {
+    public RuleName createRuleName(RuleName RuleName) {
         return ruleNameRepository.save(RuleName);
     }
 
@@ -35,7 +46,15 @@ public class RuleNameImplService implements IRuleNameService {
     }
 
     @Override
-    public void deleteRuleName(Integer id) {
-        ruleNameRepository.deleteById(id);
+    public void deleteRuleName(Integer id) throws RecordNotFoundException {
+        Optional<RuleName> ruleName = ruleNameRepository.findById(id);
+
+        if(ruleName.isPresent()) {
+            ruleNameRepository.deleteById(id);
+        } else {
+            throw new RecordNotFoundException("No record exist for given id");
+        }
+
+        //ruleNameRepository.deleteById(id);
     }
 }
